配置初始化

+++ src/main/java/makjust/utils/ResourcesInit.java
@@ -2,84 +2,86 @@ package makjust.utils;
 
 import java.io.*;
 import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.*;
+import java.util.AbstractMap;
+import java.util.List;
+import java.util.Map;
 import java.util.jar.JarFile;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
 
 public class ResourcesInit {
+    //    根路径
     private URI path = ResourcesInit.class.getProtectionDomain().getCodeSource().getLocation().toURI();
-    private String resourcesPath = path.getPath() + "/resources";
-    private List<String> checkList=Arrays.asList("config.yml","package");
+    //    目标资源地址
+    private String resourcesPath = path.getPath() + "resources";
+
     public ResourcesInit() throws Exception {
         this.mkResourcesDIR();
     }
+
     private void mkResourcesDIR() throws Exception {
         System.out.println("URI:" + path);
         if (getENV()) {
-            resourcesPath = new StringBuilder(path.getPath()).substring(0, (path.getPath().lastIndexOf("/"))) + "/resources";
+            resourcesPath = new StringBuilder(path.getPath()).substring(0, (path.getPath().lastIndexOf("/"))) + "/resources/";
             System.out.println("resourcesPath" + resourcesPath);
             copyJarResourcesFileToTemp(path, resourcesPath, "resources");
+            copyJarResourcesFileToTemp(path, resourcesPath, "plugin");
+            copyJarResourcesFileToTemp(path, resourcesPath, "package");
 
         } else {
-            checkConfig(resourcesPath);
-            copyLocalResourcesFileToTemp(new File(path.getPath()), new File(resourcesPath), true);
-
+            copyLocalResourcesFileToTemp(new File(path.getPath()), new File(resourcesPath));
         }
-
     }
 
-    //用于IDEA开发时复制resources
-    private static void copyLocalResourcesFileToTemp(File f, File nf, boolean flag) throws Exception {
+    /**
+     * 用于IDEA开发时复制resources
+     *
+     * @param f    来源
+     * @param nf   目标
+     */
+    private static void copyLocalResourcesFileToTemp(File f, File nf) throws Exception {
         // 判断是否存在
         if (f.exists()) {
-            // 判断是否是目录
             if (f.isDirectory()) {
-                if (flag) {
-                    // 制定路径，以便原样输出
-                    // 判断文件夹是否存在，不存在就创建
-                    if (!nf.exists()) {
-                        if (nf.mkdirs()){
-                            System.out.println("创建文件夹"+nf.getPath());
+                if (!nf.exists()) nf.mkdirs();
+                File[] array = f.listFiles();
+                assert array != null;
+                for (File file : array
+                ) {
+                    if (!file.isDirectory()) {
+                        if (!new File(nf.getAbsolutePath() + "/" + file.getName()).exists()) {
+                            //复制文件
+                            System.out.println("正在复制：" + file.getAbsolutePath());
+                            System.out.println("到：" + nf.getAbsolutePath() + "\\" + file.getName());
+                            // 获取输入流
+                            FileInputStream fis = new FileInputStream(file);
+                            // 获取输出流
+                            FileOutputStream fos = new FileOutputStream(nf + "/" + file.getName());
+                            byte[] b = new byte[1024];
+                            // 读取文件
+                            int len;
+                            while ((len = fis.read(b)) != -1) {
+                                // 写入文件，复制
+                                fos.write(b, 0, len);
+                            }
+                            fos.close();
+                            fis.close();
                         }
+
                     }
                 }
-                flag = true;
-                // 获取文件夹下所有的文件及子文件夹
-                File[] l = f.listFiles();
-                // 判断是否为null
-                if (null != l) {
-                    for (File ll : l) {
-                        if (ll.isDirectory()) continue;
-                        // 循环递归调用
-                        copyLocalResourcesFileToTemp(ll, nf, flag);
-                    }
-                }
-            } else {
-                System.out.println("正在复制：" + f.getAbsolutePath());
-                System.out.println("到：" + nf.getAbsolutePath() + "\\" + f.getName());
-                // 获取输入流
-                FileInputStream fis = new FileInputStream(f);
-                // 获取输出流
-                FileOutputStream fos = new FileOutputStream(nf + "/" + f.getName());
-                byte[] b = new byte[1024];
-                // 读取文件
-                int len;
-                while ((len=fis.read(b)) != -1) {
-                    // 写入文件，复制
-                    fos.write(b,0,len);
-                }
-                fos.close();
-                fis.close();
             }
         }
     }
 
-    //打jar包后复制resources
+    /**
+     * 用于打包后复制resources
+     *
+     * @param path    来源
+     * @param tempPath   目标
+     * @param filePrefix 地址
+     */
     private static void copyJarResourcesFileToTemp(URI path, String tempPath, String filePrefix) {
         try {
             List<Map.Entry<ZipEntry, InputStream>> collect =
@@ -94,6 +96,7 @@ public class ResourcesInit {
                 if (!newFile.getParentFile().exists()) {
                     newFile.getParentFile().mkdirs();
                 }
+                if (newFile.exists()) continue;
                 org.apache.commons.io.FileUtils.copyInputStreamToFile(stream, newFile);
             }
         } catch (IOException e) {
@@ -120,20 +123,9 @@ public class ResourcesInit {
         });
     }
 
-    private void checkConfig(String resourcesPath) throws IOException {
-//       File file=new File(resourcesPath);
-//        for (String fileName:checkList
-//             ) {
-//            File check=new File(resourcesPath+fileName);
-//            if (check.exists())continue;
-//            else {
-//
-//            }
-//        }
 
-    }
     //用于判断是跑idea里还是jar里
-    private  Boolean getENV() {
+    private Boolean getENV() {
         return path.getPath().contains(".jar");
     }
